requirements.txt
annotated-types==0.6.0
anyio==4.12.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
deprecation==2.1.0
exceptiongroup==1.3.1
fsspec==2025.10.0
google-auth==2.49.0
google-genai==1.46.0
gotrue==2.12.4
greenlet==3.2.5
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
playwright==1.58.0
postgrest==0.19.3
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.6.4
pydantic_core==2.16.3
pyee==13.0.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
PyPDF2==3.0.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.5.3
requests==2.32.5
resend==2.23.0
rich==14.3.3
rsa==4.9.1
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
storage3==0.11.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.11.0
supabase-auth==2.28.0
supabase-functions==2.28.0
supafunc==0.9.4
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
